refactor(pyamber): flatten over-nested util packages
Several `core/util/<X>/<X>.py` directories existed only to host a single
file — a Java/Scala "one class per directory" convention transplanted
into Python where it adds noise without value. Two related anti-patterns
existed alongside it:

- `core/util/expression_evaluator/__init__.py` was 211 lines of evaluator
  implementation (the package had no other modules).
- `core/util/virtual_identity/__init__.py` was 99 lines of helper functions.

`__init__.py` should re-export, not host the implementation.

Changes:

- Flatten single-file directories to flat modules:
  `runnable/runnable.py`, `thread/atomic.py`, `protocol/base_protocols.py`
  → `runnable.py`, `atomic.py`, `base_protocols.py`.
- Delete `core/util/operator/` — `__init__.py` was a 16-line license header
  with no exports and no importers.
- Move `expression_evaluator/__init__.py` and `virtual_identity/__init__.py`
  implementations into flat single-file modules (the dirs only ever held
  the impl + the test file).
- Move tests up alongside the new flat modules.
- Replace the hardcoded `core.util.expression_evaluator.test_expression_evaluator`
  module path inside `test_evaluate_object_expression` with `A.__module__`
  so the test tracks whatever import path it resolves to (it would have
  broken on this rename otherwise).
- Inside `core/util/stoppable/stoppable_queue_blocking_thread.py`, switch
  the intra-package `from core.util.stoppable import Stoppable` back to
  the relative `from .stoppable import Stoppable` to avoid hitting the
  parent package's __init__ mid-initialization.

Closes #4951
