+++ amber/src/main/python/core/runnables/data_processor.py
@@ -30,7 +30,7 @@ from core.models.table import all_output_to_tuple
 from core.util import Stoppable
 from core.util.console_message.replace_print import replace_print
 from core.util.console_message.timestamp import current_time_in_local_timezone
-from core.util.runnable.runnable import Runnable
+from core.util.runnable import Runnable
 from proto.org.apache.texera.amber.engine.architecture.rpc import (
     ConsoleMessage,
     ConsoleMessageType,
