+++ amber/src/main/python/core/util/atomic.py
