+++ amber/src/main/python/core/util/base_protocols.py
