+++ amber/src/main/python/core/util/buffer/buffer_base.py
@@ -17,7 +17,7 @@
 
 from abc import ABCMeta
 
-from core.util.protocol.base_protocols import FlushedGetable, Putable
+from core.util.base_protocols import FlushedGetable, Putable
 
 
 class IBuffer(FlushedGetable, Putable, metaclass=ABCMeta):
