+++ amber/src/main/python/core/util/virtual_identity.py
